corrected some errors from refactoring. updated README.
