+++ flaskr/endpoints/signup.py
@@ -1,6 +1,7 @@
 from flask import request
 
-from ..models.user import UserModel, UserSchema
+from ..models.user import UserModel
+from ..schemas.user import UserSchema
 from ..shared.authentication import Auth
 from ..shared.resource import BaseResource
